# openapi_server/openapi/openapi.yaml
openapi: 3.0.2
info:
    title: Norton
    version: 1.0.0
servers:
    -
        url: /api/v1
paths:
    /channels:
        summary: Path used to manage the list of channels.
        description: >-
            The REST endpoint/path used to list and create zero or more Channel entities.  This path contains
            a GET and POST operation to perform the list and create tasks, respectively.
        get:
            tags:
                - Channel
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Channel'
                    description: Successful response - returns an array of Channel entities.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 500
                                        messages: Unable to create user
                    description: Unable to create user
            security:
                -
                    LoginRequired: []
            operationId: get_channels
            summary: List All channels
            description: Gets a list of all Channel entities.
            x-openapi-router-controller: openapi_server.controllers.channel_controller
        post:
            requestBody:
                description: A new Channel to be created.
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Channel'
                required: true
            tags:
                - Channel
            responses:
                '201':
                    description: Successful response.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 500
                                        messages: Unable to create user
                    description: Unable to create user
            security:
                -
                    LoginRequired: []
            operationId: create_channel
            summary: Create a Channel
            description: Creates a new instance of a Channel.
            x-openapi-router-controller: openapi_server.controllers.channel_controller
    '/channels/{channelId}':
        summary: Path used to manage a single Channel.
        description: >-
            The REST endpoint/path used to get, update, and delete single instances of an Channel.  This path
            contains GET, PUT, and DELETE operations used to perform the get, update, and delete tasks,
            respectively.
        get:
            tags:
                - Channel
            parameters:
                -
                    style: simple
                    explode: false
                    name: channelId
                    description: A unique identifier for a Channel.
                    schema:
                        type: string
                    in: path
                    required: true
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Channel'
                    description: Successful response - returns a single Channel.
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 404
                                        messages: Channel not found
                    description: Could not locate Channel
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 500
                                        messages: Unable to create user
                    description: Unable to create user
            security:
                -
                    LoginRequired: []
            operationId: get_channel
            summary: Get a Channel
            description: Gets the details of a single instance of a Channel.
            x-openapi-router-controller: openapi_server.controllers.channel_controller
        put:
            requestBody:
                description: Updated Channel information.
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Channel'
                required: true
            tags:
                - Channel
            parameters:
                -
                    style: simple
                    explode: false
                    name: channelId
                    description: A unique identifier for a Channel.
                    schema:
                        type: string
                    in: path
                    required: true
            responses:
                '204':
                    description: Channel has been updated.
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 404
                                        messages: Channel not found
                    description: Could not locate Channel
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 500
                                        messages: Unable to create user
                    description: Unable to create user
            security:
                -
                    LoginRequired: []
            operationId: update_channel
            summary: Update a Channel
            description: Updates an existing Channel.
            x-openapi-router-controller: openapi_server.controllers.channel_controller
        delete:
            tags:
                - Channel
            parameters:
                -
                    style: simple
                    explode: false
                    name: channelId
                    description: A unique identifier for a Channel.
                    schema:
                        type: string
                    in: path
                    required: true
            responses:
                '204':
                    description: Successful response.
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 404
                                        messages: Channel not found
                    description: Could not locate Channel
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 500
                                        messages: Unable to create user
                    description: Unable to create user
            security:
                -
                    LoginRequired: []
            operationId: delete_channel
            summary: Delete a Channel
            description: Deletes an existing Channel.
            x-openapi-router-controller: openapi_server.controllers.channel_controller
    '/channels/{channelId}/members':
        summary: Path used to manage the all Users in a Channel
        get:
            tags:
                - Channel
            parameters:
                -
                    style: simple
                    explode: false
                    name: channelId
                    schema:
                        type: integer
                    in: path
                    required: true
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/User'
                            examples:
                                SampleUsers:
                                    value:
                                        -
                                            id: 1
                                            username: David
                                        -
                                            id: 2
                                            username: Morgan
                    description: Returns a list of Users for the specified Channel
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 404
                                        messages: Channel not found
                    description: Could not locate Channel
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 500
                                        messages: Unable to create user
                    description: Unable to create user
            security:
                -
                    LoginRequired: []
            operationId: get_channel_members
            summary: List All members
            description: Gets a list of all Users in a given Channel.
            x-openapi-router-controller: openapi_server.controllers.channel_controller
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/User'
                required: true
            tags:
                - Channel
            parameters:
                -
                    style: simple
                    explode: false
                    name: channelId
                    schema:
                        type: integer
                    in: path
                    required: true
            responses:
                '201':
                    description: User has been created
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 404
                                        messages: Channel not found
                    description: Could not locate Channel
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 500
                                        messages: Unable to create user
                    description: Unable to create user
            operationId: add_user_to_channel
            x-openapi-router-controller: openapi_server.controllers.channel_controller
    '/channels/{channelId}/members/{userId}':
        summary: Path used to manage the indidual Users in a Channel
        delete:
            tags:
                - Channel
            parameters:
                -
                    style: simple
                    explode: false
                    name: channelId
                    schema:
                        type: integer
                    in: path
                    required: true
                -
                    style: simple
                    explode: false
                    name: userId
                    schema:
                        type: integer
                    in: path
                    required: true
            responses:
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 404
                                        messages: Channel not found
                    description: Could not locate Channel
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 500
                                        messages: Unable to create user
                    description: Unable to create user
            security:
                -
                    LoginRequired: []
            operationId: remove_channel_member
            summary: Remove as User
            description: Removes a User for a Channel
            x-openapi-router-controller: openapi_server.controllers.channel_controller
    '/channels/{channelId}/messages':
        summary: Path used to manage the list of messages.
        description: >-
            The REST endpoint/path used to list and create zero or more Message entities.  This path contains
            a GET and POST operation to perform the list and create tasks, respectively.
        get:
            tags:
                - Channel
            parameters:
                -
                    style: simple
                    explode: false
                    name: channelId
                    schema:
                        type: integer
                    in: path
                    required: true
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Message'
                    description: Successful response - returns an array of Message entities.
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 404
                                        messages: Channel not found
                    description: Could not locate Channel
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 500
                                        messages: Unable to create user
                    description: Unable to create user
            security:
                -
                    LoginRequired: []
            operationId: get_messages
            summary: List All messages
            description: Gets a list of all Message entities.
            x-openapi-router-controller: openapi_server.controllers.channel_controller
        post:
            requestBody:
                description: A new Message to be created.
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Message'
                required: true
            tags:
                - Channel
            parameters:
                -
                    style: simple
                    explode: false
                    name: channelId
                    schema:
                        type: integer
                    in: path
                    required: true
            responses:
                '201':
                    description: Successful response.
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 404
                                        messages: Channel not found
                    description: Could not locate Channel
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                ChannelNotFound:
                                    value:
                                        code: 500
                                        messages: Unable to create user
                    description: Unable to create user
            security:
                -
                    LoginRequired: []
            operationId: create_message
            summary: Create a Message
            description: Creates a new instance of a Message.
            x-openapi-router-controller: openapi_server.controllers.channel_controller
...
components:
    schemas:
        Channel:
            required:
                - id
                - name
                - type
            type: object
            properties:
                id:
                    type: integer
                name:
                    type: string
                type:
                    enum:
                        - TEXT
                        - VOICE
                    type: string
            example:
                name: name
                id: 0
                type: TEXT
        User:
            required:
                - id
                - name
            type: object
            properties:
                id:
                    type: integer
                name:
                    type: string
            example:
                id: 0
                name: name
        Message:
            required:
                - author
                - channel
                - content
                - timestamp
            type: object
            properties:
                id:
                    type: integer
                channel:
                    type: integer
                author:
                    type: integer
                content:
                    type: string
                timestamp:
                    type: integer
            example:
                author: author
                channel: channel
                id: 0
                content: content
                timestamp: 6
        Error:
            required:
                - code
                - messages
            type: object
            properties:
                code:
                    type: integer
                messages:
                    type: string
...
    securitySchemes:
        LoginRequired:
            type: apiKey
            name: session_id
            in: cookie
            x-apikeyInfoFunc: openapi_server.controllers.security_controller_.info_from_LoginRequired